洗车场收入列表、评价、更新营业状态、发布/删除洗车服务、增加/修改/删除商品

+++ iOS Code/BubbleCarWash/BubbleCarWash/Home/SubViews/CarWash/Model/CarWashServiceModel.h
@@ -6,7 +6,7 @@
 //  Copyright © 2018年 Sunshine Girl. All rights reserved.
 //
 
-#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
 
 typedef void(^ServiceResultBlock)(NSArray *result);
 
@@ -20,7 +20,7 @@ typedef void(^ServiceResultBlock)(NSArray *result);
 
 @property (nonatomic, copy) NSString *carWashName;
 @property (nonatomic, copy) NSString *desc;
-@property (nonatomic, assign) NSInteger price;
+@property (nonatomic, assign) CGFloat price;
 @property (nonatomic, assign) NSInteger dataID;
 
 - (instancetype)initWithDic:(NSDictionary *)dic;
